backend/requirements.txt
# --- PyTorch GPU (Windows/CUDA 12.4) ---
# Ensuring GPU versions are preferred during installation
# --- GPU Engines (Already installed manually) ---
torch
torchaudio

# --- Core API ---
numpy<2.0.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.0.0
python-multipart>=0.0.12
python-dotenv>=1.0.0

# --- Database ---
sqlalchemy>=2.0.0
aiosqlite>=0.20.0

# --- GPU Monitoring & Support ---
# official replacement for pynvml
nvidia-ml-py>=12.535.133

# --- Inference Engines ---
onnxruntime-gpu>=1.20.0
transformers>=4.44.0
accelerate>=0.33.0
bitsandbytes>=0.43.0
pyannote.audio>=3.1.0

# --- Audio I/O ---
soundfile>=0.12.1

# --- Testing ---
pytest>=7.4.0
httpx>=0.25.0
pytest-asyncio>=0.23.0
